+++ hisyonosuke/src/birthday-registrator/notificator.ts
